+++ cw-orch-daemon/src/state.rs
@@ -1,56 +1,67 @@
 use super::error::DaemonError;
-use crate::{
-    channel::GrpcChannel, env::default_state_folder, env::DaemonEnvVars, networks::ChainKind,
-};
+use crate::env::{default_state_folder, DaemonEnvVars};
+use crate::{json_lock::JsonLockedState, networks::ChainKind};
 
 use cosmwasm_std::Addr;
 use cw_orch_core::environment::ChainInfoOwned;
-use cw_orch_core::{
-    environment::StateInterface,
-    log::{connectivity_target, local_target},
-    CwEnvError,
-};
+use cw_orch_core::{environment::StateInterface, log::local_target, CwEnvError};
+use once_cell::sync::Lazy;
 use serde::Serialize;
 use serde_json::{json, Value};
-use std::{collections::HashMap, fs::File, path::Path};
-use tonic::transport::Channel;
+use std::sync::Arc;
+use std::{
+    collections::{HashMap, HashSet},
+    path::Path,
+    sync::Mutex,
+};
+
+/// Global state to track which files are already open by other daemons from other threads
+/// This is necessary because File lock will allow same process to lock file how many times as process wants
+pub(crate) static LOCKED_FILES: Lazy<Mutex<HashSet<String>>> =
+    Lazy::new(|| Mutex::new(HashSet::new()));
 
 /// Stores the chain information and deployment state.
 /// Uses a simple JSON file to store the deployment information locally.
-#[derive(Clone, Debug)]
+#[derive(Debug, Clone)]
 pub struct DaemonState {
-    /// this is passed via env var STATE_FILE
-    pub json_file_path: String,
+    pub json_state: DaemonStateFile,
     /// Deployment identifier
     pub deployment_id: String,
-    /// gRPC channel
-    pub grpc_channel: Channel,
     /// Information about the chain
     pub chain_data: ChainInfoOwned,
-    /// Flag to set the daemon state readonly and not pollute the env file
-    pub read_only: bool,
+}
+
+impl Drop for DaemonState {
+    fn drop(&mut self) {
+        if let DaemonStateFile::FullAccess { json_file_state } = &self.json_state {
+            let json_lock = json_file_state.lock().unwrap();
+            let mut locked_files = LOCKED_FILES.lock().unwrap();
+            locked_files.remove(json_lock.path());
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum DaemonStateFile {
+    ReadOnly {
+        path: String,
+    },
+    FullAccess {
+        json_file_state: Arc<Mutex<JsonLockedState>>,
+    },
 }
 
 impl DaemonState {
     /// Creates a new state from the given chain data and deployment id.
     /// Attempts to connect to any of the provided gRPC endpoints.
-    pub async fn new(
+    pub fn new(
+        mut json_file_path: String,
         chain_data: ChainInfoOwned,
         deployment_id: String,
         read_only: bool,
     ) -> Result<DaemonState, DaemonError> {
-        if chain_data.grpc_urls.is_empty() {
-            return Err(DaemonError::GRPCListIsEmpty);
-        }
-
-        log::debug!(target: &connectivity_target(), "Found {} gRPC endpoints", chain_data.grpc_urls.len());
-
-        // find working grpc channel
-        let grpc_channel =
-            GrpcChannel::connect(&chain_data.grpc_urls, chain_data.chain_id.as_str()).await?;
-
-        // If the path is relative, we dis-ambiguate it and take the root at $HOME/$CW_ORCH_STATE_FOLDER
-        let mut json_file_path = Self::state_file_path()?;
+        let chain_id = &chain_data.chain_id;
+        let chain_name = &chain_data.network_info.chain_name;
 
         log::debug!(target: &local_target(), "Using state file : {}", json_file_path);
 
@@ -70,33 +81,36 @@ impl DaemonState {
             json_file_path = format!("{folder}/{name}_local.json");
         }
 
-        // build daemon state
-        let state = DaemonState {
-            json_file_path,
-            deployment_id,
-            grpc_channel,
-            chain_data,
-            read_only,
-        };
-
-        if !read_only {
+        let json_state = if read_only {
+            DaemonStateFile::ReadOnly {
+                path: json_file_path,
+            }
+        } else {
             log::info!(
                 target: &local_target(),
-                "Writing daemon state JSON file: {:#?}",
-                state.json_file_path
+                "Writing daemon state JSON file: {json_file_path:#?}",
             );
 
-            // write json state file
-            crate::json_file::write(
-                &state.json_file_path,
-                &state.chain_data.chain_id.to_string(),
-                &state.chain_data.network_info.chain_name.to_string(),
-                &state.deployment_id,
-            );
-        }
+            let mut lock = LOCKED_FILES.lock().unwrap();
+            if lock.contains(&json_file_path) {
+                return Err(DaemonError::StateAlreadyLocked(json_file_path));
+            }
+            let mut json_file_state = JsonLockedState::new(&json_file_path);
+            // Insert and drop mutex lock asap
+            lock.insert(json_file_path);
+            drop(lock);
 
-        // finish
-        Ok(state)
+            json_file_state.prepare(chain_id, chain_name, &deployment_id);
+            DaemonStateFile::FullAccess {
+                json_file_state: Arc::new(Mutex::new(json_file_state)),
+            }
+        };
+
+        Ok(DaemonState {
+            json_state,
+            deployment_id,
+            chain_data,
+        })
     }
 
     /// Returns the path of the file where the state of `cw-orchestrator` is stored.
@@ -133,38 +147,83 @@ impl DaemonState {
 
         Ok(state_file_path)
     }
-    /// Get the state filepath and read it as json
-    fn read_state(&self) -> Result<serde_json::Value, DaemonError> {
-        crate::json_file::read(&self.json_file_path)
-    }
 
     /// Retrieve a stateful value using the chainId and networkId
     pub fn get(&self, key: &str) -> Result<Value, DaemonError> {
-        let json = self.read_state()?;
-        Ok(
-            json[&self.chain_data.network_info.chain_name][&self.chain_data.chain_id.to_string()]
-                [key]
+        let json = match &self.json_state {
+            DaemonStateFile::ReadOnly { path } => {
+                let j = crate::json_lock::read(path)?;
+
+                j[&self.chain_data.network_info.chain_name][&self.chain_data.chain_id].clone()
+            }
+            DaemonStateFile::FullAccess { json_file_state } => json_file_state
+                .lock()
+                .unwrap()
+                .get(
+                    &self.chain_data.network_info.chain_name,
+                    &self.chain_data.chain_id,
+                )
                 .clone(),
-        )
+        };
+        Ok(json[key].clone())
     }
 
     /// Set a stateful value using the chainId and networkId
     pub fn set<T: Serialize>(
-        &self,
+        &mut self,
         key: &str,
         contract_id: &str,
         value: T,
     ) -> Result<(), DaemonError> {
-        if self.read_only {
-            return Err(DaemonError::StateReadOnly);
+        let json_file_state = match &mut self.json_state {
+            DaemonStateFile::ReadOnly { path } => {
+                return Err(DaemonError::StateReadOnly(path.clone()))
+            }
+            DaemonStateFile::FullAccess { json_file_state } => json_file_state,
+        };
+
+        let mut json_file_lock = json_file_state.lock().unwrap();
+        let val = json_file_lock.get_mut(
+            &self.chain_data.network_info.chain_name,
+            &self.chain_data.chain_id,
+        );
+        val[key][contract_id] = json!(value);
+
+        Ok(())
+    }
+
+    /// Forcefully write current json to a file
+    pub fn force_write(&mut self) -> Result<(), DaemonError> {
+        let json_file_state = match &mut self.json_state {
+            DaemonStateFile::ReadOnly { path } => {
+                return Err(DaemonError::StateReadOnly(path.clone()))
+            }
+            DaemonStateFile::FullAccess { json_file_state } => json_file_state,
+        };
+        json_file_state.lock().unwrap().force_write();
+        Ok(())
+    }
+
+    /// Flushes all the state related to the current chain
+    /// Only works on Local networks
+    pub fn flush(&mut self) -> Result<(), DaemonError> {
+        if self.chain_data.kind != ChainKind::Local {
+            panic!("Can only flush local chain state");
         }
+        let json_file_state = match &mut self.json_state {
+            DaemonStateFile::ReadOnly { path } => {
+                return Err(DaemonError::StateReadOnly(path.clone()))
+            }
+            DaemonStateFile::FullAccess { json_file_state } => json_file_state,
+        };
 
-        let mut json = self.read_state()?;
+        let mut lock = json_file_state.lock().unwrap();
+        let json = lock.get_mut(
+            &self.chain_data.network_info.chain_name,
+            &self.chain_data.chain_id,
+        );
 
-        json[&self.chain_data.network_info.chain_name][&self.chain_data.chain_id.to_string()]
-            [key][contract_id] = json!(value);
-
-        serde_json::to_writer_pretty(File::create(&self.json_file_path).unwrap(), &json)?;
+        *json = json!({});
         Ok(())
     }
 }
@@ -173,8 +232,9 @@ impl StateInterface for DaemonState {
     /// Read address for contract in deployment id from state file
     fn get_address(&self, contract_id: &str) -> Result<Addr, CwEnvError> {
         let value = self
-            .get(&self.deployment_id)?
-            .get(contract_id)
+            .get(&self.deployment_id)
+            .ok()
+            .and_then(|v| v.get(contract_id).cloned())
             .ok_or_else(|| CwEnvError::AddrNotInStore(contract_id.to_owned()))?
             .clone();
         Ok(Addr::unchecked(value.as_str().unwrap()))
@@ -182,15 +242,17 @@ impl StateInterface for DaemonState {
 
     /// Set address for contract in deployment id in state file
     fn set_address(&mut self, contract_id: &str, address: &Addr) {
-        self.set(&self.deployment_id, contract_id, address.as_str())
+        let deployment_id = self.deployment_id.clone();
+        self.set(&deployment_id, contract_id, address.as_str())
             .unwrap();
     }
 
     /// Get the locally-saved version of the contract's version on this network
     fn get_code_id(&self, contract_id: &str) -> Result<u64, CwEnvError> {
         let value = self
-            .get("code_ids")?
-            .get(contract_id)
+            .get("code_ids")
+            .ok()
+            .and_then(|v| v.get(contract_id).cloned())
             .ok_or_else(|| CwEnvError::CodeIdNotInStore(contract_id.to_owned()))?
             .clone();
         Ok(value.as_u64().unwrap())
@@ -276,6 +338,7 @@ pub mod test {
             parent_and_relative_state_path
         );
 
+        std::env::remove_var(STATE_FILE_ENV_NAME);
         Ok(())
     }
 }