+++ cw-orch-daemon/tests/index.rs
@@ -15,16 +15,10 @@ mod tests {
             .build()
             .unwrap();
 
-        let indexed_daemon = Daemon::builder()
-            .chain(networks::LOCAL_JUNO)
-            .hd_index(56)
-            .build()
-            .unwrap();
+        let daemon_sender = daemon.sender().to_string();
+        let indexed_daemon = daemon.rebuild().hd_index(56).build().unwrap();
 
-        assert_ne!(
-            daemon.sender().to_string(),
-            indexed_daemon.sender().to_string()
-        );
+        assert_ne!(daemon_sender, indexed_daemon.sender().to_string());
 
         Ok(())
     }