+++ cw-orch-interchain/tests/common/mod.rs
@@ -1,0 +1,2 @@
+pub mod bank;
+pub mod ica_demo;